nsqlookupd: producer tombstones
when a topic producer goes away (operationally moved to another host,
etc.) it isn't possible to just delete the topic using the existing
functionality because you only want to delete the topic from a single
host.  also, because clients connect to all producers if you try to
issue an HTTP request directly to that (old) producer to delete the
topic there are race conditions which prevent that from working
correctly.

this change adds the ability to "tombstone" a specific producer of a
topic so that lookup queries from clients will no longer receive that
producer in their response, allowing you to then issue a command
directly to that nsqd to delete the topic.