nsqd: fix new topic channel lookup/creation
when an nsqd first sees a topic queries nsqlookupd to identify
any channels that the cluster might know about in order
to proactively create them (in the case where there is more than one
channel this prevents data loss due to timing of consumers subscribing).

in #176 a regression was introduced that prevent the topic's messagePump
state from being updated in the code path described above.  This meant
that an nsqd that identified channels during creation of a topic would
not begin writing those messages into its channels.