nsqd: fix race for client subscription cleanup
SubEventChan is buffered to prevent deadlocks so it
isn't guaranteed that messagePump() will recv
the channel subscription state update.

If a client races and disconnects, messagePump()
exit will not tell the channel to remove the
client metadata.

This moves the code to the readLoop() exit path
which has no memory race reading client.Channel
and can always properly cleanup.