allow <SHORT_HOST> to be used in the filename
Allow <SHORT_HOST> to be used in the file name.  This will help facilitate creating unique file names if the output of multiple nsq_to_file processes on different machines are to be stored in a single consolidated location.  The <HOST> argument provides the same benefit but <SHORT_HOST> is, well, shorter and a little cleaner.