Create Dockerfiles.
Create a base Dockerfile that builds nsq and all its subdirectories.

Create a Dockerfile for nsqd that inherits the base and will run nsqd
with the passed command line flags.

Create a Dockerfile for nsqlookupd that inherits the base and will run
nsqlookupd with the passed command line flags.

Create a Dockerfile for nsqadmin that inherits the base and will run
nsqadmin with the passed command line flags.