Second attempt at a Docker strategy.
Now we're trying to build the smallest container possible, and only ship
one container, instead of one container per app.

The strategy is to create a `docker` target in `dist`, and add all the
binaries from the linux target to that folder. The Dockerfile is based
off of `scratch`, and adds the `docker` target in `dist` at the root of
the filesystem.

This gives us a Docker image with about 5MB of overhead over the
binaries, compared to our previous strategy's ~700MB. It also means that
the versions will be easier to maintain and control.

The resulting images can be used as follows:

`docker run nsqio/nsq /nsqd --version`

Where `nsqd` is the binary you want to run. Unfortunately, I believe `/`
is required.

Remove old Dockerfiles and instructions.

Remove the old Dockerfiles and the old Docker instructions from the
binary readmes.