Enable shell expansion.
We're changing from scratch to busybox as a base image, after discussion
in #543. This gives us some free shell expansion.

I also added a step at the end of dist.sh that would take care of
tagging the image as the "latest" build, as long as there's not a hyphen
in the version number (so our "latest" build will never be an alpha or
beta, as I assume we want stability by default).