nsqd: do not print EOF when closing cleanly.
When nsqd shuts down, protocol_v2 responds to CLS with CLOSE_WAIT and then the
IOLoop waits for another command with ReadSlice on the buffered client
connection. ReadSlice could return an error because the client closes the
connection or if the read buffer is full.  This error is returned from IOLoop
to the tcp handler and printed.

Because we are shutting down, we do not care about the error.