nsq_to_{nsq,http}: add epsilon-greedy hostpool mode
go-hostpool's EpsilonGreedy chooses hosts to send to based on which host has
been most performant in the past, which is a good behavior to add.

The current hostpool option round-robins with backoff on failing hosts. Because
the current hostpool is a better round-robin, we are switching the default mode
to hostpool.