nsqd: Ensure we print an error if a bad ID is generated
Because lastError is initialised to time.Now(), a full second must
elapse before we'll print an error. If an error is returned at least
once a second after that, we'll never print an error.

Instead we can initialise lastError to the UNIX Epoch timestamp, meaning
the first error we see will generate an error message.