refactor code organisation of guid, message and queue
first, we make the GUID library publically usable.
This requires a few other needed changes that are also useful
in their own right, basically make small packages unaware
of the things that implement them.
I.e. GUID's need not be aware of the messages using them, and
messages need not be aware of the queues using them.