nsqadmin: fix js source-map serving for firefox
The Content-Type of the source-map file is not supposed to matter.
But, if Content-Encoding is gzip, and Content-Type is x-gzip, then
Firefox does not decompress the response before trying to decode json,
and fails to load the source map for the debugger, complaining:
"unexpected character at line 1 column 1 of the JSON data".
With a proper Content-Type, Firefox decompresses and understands.
(Chrome works either way.)