nsqadmin: add option to skip resolving on startup
When running nsqlookupd in Kubernetes and pointing to a specific
identity of a pod from statefulset, the DNS name is not resolvable
if the pod is not currently scheduled.

The lookup check currently causes nsqadmin to exit during startup
even if only one of the listed nsqlookupd instances is not resolvable.

I want nsqadmin to always start. If no lookupd instances are currently
resolvable/reachable that fact will be visible in the user interface.