nsqd: allow disable both HTTP and HTTPS interfaces
NSQD by defaults opens both the HTTP and well as the HTTPS interface
by default (if TLS is set). This is fine as a default behavior, but
it's valuable to be able to disable one interface or another to raise
the security guarantees (e.g. if TCP is enough, there is no reason to
keep 2 more ports open just to have to lock them down in some other
way).

This commits allows disabling the HTTP interface the same way the HTTPS
one could be disabled until now (set it to ""). In addition, it makes
a few code cleanups.