nsqd: add memory release function
In some cases where the message volume is large, the memory occupied by the program cannot be returned to the operating system immediately. You can call this method to enforce execution. However, using this method may cause the system to pause for a period of time, so use it with caution.